+++ public_api.ts
@@ -173,6 +173,7 @@ export * from './src/app/modules/lux-form/lux-file/lux-file-model/lux-file-objec
 // Directives
 export * from './src/app/modules/lux-form/lux-file/lux-file-model/lux-file-capture.directive';
 export * from './src/app/modules/lux-form/lux-form-control/lux-form-directives/lux-maxlength/lux-max-length.directive';
+export * from './src/app/modules/lux-form/lux-form-control/lux-form-directives/lux-name/lux-name-directive.directive';
 
 /**
  *   _    _   ___  __     _____ ___ _     _____      ____  ____  _______     _____ _______        __
@@ -353,3 +354,31 @@ export * from './src/app/modules/lux-util/lux-media-query-observer.service';
 export * from './src/app/modules/lux-util/testing/lux-test-helper';
 export * from './src/app/modules/lux-util/testing/lux-test-overlay-helper';
 export * from './src/app/modules/lux-util/lux-paginator-intl';
+
+/**
+ *  _    _   ___  __     _   _ _____ __  __ _
+ * | |  | | | \ \/ /    | | | |_   _|  \/  | |
+ * | |  | | | |\  /_____| |_| | | | | |\/| | |
+ * | |__| |_| |/  \_____|  _  | | | | |  | | |___
+ * |_____\___//_/\_\    |_| |_| |_| |_|  |_|_____|
+ *
+ */
+export * from './src/app/modules/lux-html/lux-html.module';
+// Components
+export * from './src/app/modules/lux-html/lux-html/lux-html.component';
+// Pipes
+export * from './src/app/modules/lux-html/lux-sanitize/lux-sanitize.pipe';
+// Classes
+export * from './src/app/modules/lux-html/lux-sanitize/lux-sanitize-config';
+
+/**
+ *  _    _   ___  __     __  __    _    ____  _  ______   _____        ___   _
+ * | |  | | | \ \/ /    |  \/  |  / \  |  _ \| |/ /  _ \ / _ \ \      / / \ | |
+ * | |  | | | |\  /_____| |\/| | / _ \ | |_) | ' /| | | | | | \ \ /\ / /|  \| |
+ * | |__| |_| |/  \_____| |  | |/ ___ \|  _ <| . \| |_| | |_| |\ V  V / | |\  |
+ * |_____\___//_/\_\    |_|  |_/_/   \_\_| \_\_|\_\____/ \___/  \_/\_/  |_| \_|
+ *
+ */
+export * from './src/app/modules/lux-markdown/lux-markdown.module';
+// Components
+export * from './src/app/modules/lux-markdown/lux-markdown/lux-markdown.component';