+++ src/app/demo/components-overview/components-overview-navigation.service.ts
@@ -95,7 +95,9 @@ export class ComponentsOverviewNavigationService {
       icon: 'fas fa-bell',
       label: 'Badge-Notification',
       moduleName: 'directive'
-    }
+    },
+    { onclick: () => this.goTo('html'), icon: 'fab fa-html5', label: 'Html', moduleName: 'html' },
+    { onclick: () => this.goTo('markdown'), icon: 'fab fa-markdown', label: 'Markdown', moduleName: 'markdown' }
   ];
 
   sortedComponentEntries: Map<string, any[]> = new Map<string, any[]>();
@@ -105,9 +107,11 @@ export class ComponentsOverviewNavigationService {
     ['directive', false],
     ['error', false],
     ['form', false],
+    ['html', false],
     ['icon', false],
     ['layout', false],
     ['lookup', false],
+    ['markdown', false],
     ['pipes', false],
     ['popup', false]
   ]);